requirements.txt
discord.py==2.0.0
aiohttp==3.8.1
xlsxwriter==3.0.2
